TEOTB-6: Update precisions for some fields

+++ TelegramBotCore/src/main/java/com/github/donkeyrit/models/response/Audio.java
@@ -37,13 +37,10 @@ public record Audio(
      */
     Optional<String> mimeType,
     /**
-     * File size in bytes. It can be bigger than 2^31 and some programming languages 
-     * may have difficulty/silent defects in interpreting it. 
-     * But it has at most 52 significant bits, so a signed 64-bit integer or double-precision float type 
-     * are safe for storing this value.
+     * File size in bytes. 
+     * It can be bigger than 2^31.
      */
-    //TODO: Update precision
-    Optional<Integer> fileSize,
+    Optional<Double> fileSize,
     /**
      * Thumbnail of the album cover to which the music file belongs
      */
